profai/V1/requirements.txt
openai==1.3.8
python-dotenv==1.0.0
rich==13.7.0